refactor (nextjs): rename component
