+++ nextjs/src/components/cards/card-slider.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import type { Card } from "./card"
 import Text from "../text"
 import CardColored from "./card-colored"
-import Link from "../link"
+import Link from "../link-button"
 import CardSliderExplanation from "./card-slider-explanation"
 import type { CTA } from "@/lib/cta"
 import ButtonGroup from "../button-group"
